requirements.txt
timm
matplotlib
pandas
numpy
opencv-python
onnx
onnxruntime
carla
pygame
numpy
pycocotools
eclipse-zenoh